+++ src/model_config.h
@@ -12,6 +12,7 @@ enum NBFC_PACKED_ENUM RegisterWriteMode_ {
   RegisterWriteMode_And,
   RegisterWriteMode_Or,
   RegisterWriteMode_Call,
+  RegisterWriteMode_Lua,
   RegisterWriteMode_Unset,
 };